Revert use of native microtime
This was a bad mistake on my part. `process.hrtime` is not a unix time;
rather it's relative to 'an arbitrary time in the past'. It therefore
cannot be used as a rate limiter key, since it's not consistent across
sessions.

I reverted this change, although I updated to the newer version of the
microtime library in the process.